# resources/api/openapi.yml
openapi: 3.0.3
info:
  version: 1.0.0-beta
  title: CritterSystem API
  description: >
    The CritterSystem API provides comprehensive access to event management functionality.
    This API is as stable as a **beta** version might be.
    It could burst into flames and morph into a monster at any second!
    (But we try to keep it somewhat consistent, at least during events).
  contact:
    name: GitHub Issues
    url: https://github.com/engelsystem/engelsystem/issues
  license:
    name: GPL 2.0
    # identifier: GPL-2.0
servers:
  - url: /api/v0-beta
    description: This server (Legacy v0-beta)
  - url: /api/v2
    description: This server (v2)
  - url: '{host}/api/v0-beta'
    description: CritterSystem instance to use (Legacy v0-beta)
    variables:
      host:
        default: http://localhost:5080
  - url: '{host}/api/v2'
    description: CritterSystem instance to use (v2)
    variables:
      host:
        default: http://localhost:5080
...
security:
  - bearer-auth: [ ]
  - api-key-header: [ ]
components:
  securitySchemes:
    bearer-auth:
      type: http
      scheme: bearer
      bearerFormat: API key from settings
    api-key-header:
      type: apiKey
      name: x-api-key
      in: header
  responses:
    UnauthorizedError: # 401
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    UserCertification:
      type: object
      properties:
        id:
          type: integer
          example: 42
        user:
          $ref: '#/components/schemas/Reference'
        certification:
          $ref: '#/components/schemas/Reference'
        status:
          type: string
          enum: [active, expired, revoked, pending, applied]
          example: active
        issued_at:
          $ref: '#/components/schemas/DateTimeOptional'
        expires_at:
          $ref: '#/components/schemas/DateTimeOptional'
        revoked_at:
          $ref: '#/components/schemas/DateTimeOptional'
        notes:
          type: string
          nullable: true
          example: Completed advanced training module
        verified_by:
          $ref: '#/components/schemas/Reference'
        created_at:
          $ref: '#/components/schemas/DateTime'
        updated_at:
          $ref: '#/components/schemas/DateTimeOptional'
      required:
        - id
        - user
        - certification
        - status
        - created_at
        - updated_at
...
    DateTime:
      type: string
      format: date-time
      description: DateTime in ISO-8601 format
      example: 2023-05-23T00:00:00.000000Z
    DateTimeOptional:
      type: string
      format: date-time
      description: DateTime in ISO-8601 format
      example: 2023-05-23T00:00:00.000000Z
      nullable: true
    Error:
      type: object
      properties:
        message:
          type: string
...
    Reference:
      type: object
      description: Reference to other resources
      properties:
        id:
          type: integer
          example: 42
          description: The ID of the related resource
        name:
          type: string
          nullable: true
          example: Foo
          description: A name of the related resource, should only be used for reference
      required:
        - id
paths:
...
  /user/certifications:
    get:
      tags:
        - certification
        - user
      summary: Get user's certifications
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserCertification'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...